pyproject.toml
[tool.poetry]
name = "python-django-diploma"
version = "0.1.0"
description = ""
authors = ["admin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.0"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-debug-toolbar = "^4.1.0"
pillow = "^9.5.0"
gunicorn = "^20.1.0"
diploma-frontend = "0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
